Pass tests

+++ src/cli/cli-cfunc.c
@@ -93,7 +93,7 @@ static bool serialize_printer(
 	int ival;
 	float fval;
 	const char *sval;
-	uint32_t items, i;
+	size_t items, i;
 	char digits[1024];
 
 	if (!noct_get_value_type(env, value, &type))
