+++ src/core/gc.h
@@ -199,7 +199,7 @@ struct rt_dict *rt_gc_alloc_dict(struct rt_env *env, size_t size);
 struct rt_packed *rt_gc_alloc_packed(struct rt_env *env, int type, size_t size, size_t elem_size, void *preallocated);
 
 /* Write barrier: registers a container in the remember set if it references a young object. */
-void rt_gc_array_write_barrier(struct rt_env *env, struct rt_array *arr, uint32_t index, struct rt_value *val);
+void rt_gc_array_write_barrier(struct rt_env *env, struct rt_array *arr, size_t index, struct rt_value *val);
 
 /* Write barrier: registers a container in the remember set if it references a young object. */
 void rt_gc_dict_write_barrier(struct rt_env *env, struct rt_dict *dict, struct rt_value *val);
